+++ node/clients/index.ts
@@ -1,5 +1,6 @@
 import { ClientsConfig, IOClients, LRUCache } from '@vtex/api'
 
+import Sitemap from './sitemap'
 import WordpressProxyDataSource from './wordpressProxy'
 
 const TIMEOUT_MS = 8000
@@ -16,6 +17,10 @@ metrics.trackCache('wordpressProxy', cacheStorage)
 export class Clients extends IOClients {
   public get wordpressProxy() {
     return this.getOrSet('wordpressProxy', WordpressProxyDataSource)
+  }
+
+  public get sitemap() {
+    return this.getOrSet('sitemap', Sitemap)
   }
 }